test: 함수 위치 이동에 따른 테스트 클래스 분리
